VTO/virtual-try-on/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
pillow
python-multipart
